refactor: reduce `W32` compilation units
